movimentação corrigida: variáveis que deviam ser FLOAT estavam como INT sprites DOWN, LEFT, RIGHT, UP, incluidos, mas ainda nao implementados função movimento do pacman reescrita no loop do game

+++ pacman-sin141/entidade.cpp
@@ -2,10 +2,10 @@
 #include <allegro5/allegro5.h>
 #include <allegro5/allegro_image.h>
 
-int Entidade::getX() {
+float Entidade::getX() {
 	return this->posX;
 }
-int Entidade::getY() {
+float Entidade::getY() {
 	return this->posY;
 }
 void Entidade::setX(float x) {
